% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Nimble Management.R
\name{nimbleManagement}
\alias{nimbleManagement}
\title{Nimble management visualization}
\usage{
nimbleManagement(
  N,
  allToAll = FALSE,
  nams = c("Problem", "Data", "Models", "Analyses", "Forecast", "Risk Assessment",
    "Decisions"),
  cols = "RdYlBu"
)
}
\description{
Nimble management visualization
}
\examples{
if (require(animation) && require(quickPlot)) {
dev(width = 11, height = 11)
nimbleManagement(7, cols = "Set2", allToAll = TRUE)
saveGIF(movie.name = "Nimble Management.gif", 
        nimbleManagement(7, cols = "Accent", allToAll = TRUE),
        ani.options = ani.options(
          ani.height = 1200, 
          ani.width = 1200, ani.res = 200), {
            
          })
}
}
